Extracting flag struct
