+++ pkg/cmd/cmd.go
@@ -17,7 +17,11 @@
 package cmd
 
 import (
+	"fmt"
+	"reflect"
+
 	"github.com/spf13/cobra"
+	"github.com/spf13/pflag"
 
 	"github.com/ZupIT/ritchie-cli/pkg/api"
 	"github.com/ZupIT/ritchie-cli/pkg/prompt"
@@ -26,8 +30,34 @@ import (
 const stdinWarning = "stdin commands are deprecated and will no longer be supported in future versions. Please use" +
 	" flags for programatic formula execution"
 
+type flag struct {
+	name        string
+	shortName   string
+	kind        reflect.Kind
+	defValue    interface{}
+	description string
+}
+
+type flags []flag
+
 // CommandRunnerFunc represents that runner func for commands.
 type CommandRunnerFunc func(cmd *cobra.Command, args []string) error
+
+func addReservedFlags(flags *pflag.FlagSet, flagsToAdd flags) {
+	for _, flag := range flagsToAdd {
+		switch flag.kind { //nolint:exhaustive
+		case reflect.String:
+			flags.StringP(flag.name, flag.shortName, flag.defValue.(string), flag.description)
+		case reflect.Bool:
+			flags.BoolP(flag.name, flag.shortName, flag.defValue.(bool), flag.description)
+		case reflect.Int:
+			flags.IntP(flag.name, flag.shortName, flag.defValue.(int), flag.description)
+		default:
+			warning := fmt.Sprintf("The %q type is not supported for the %q flag", flag.kind.String(), flag.name)
+			prompt.Warning(warning)
+		}
+	}
+}
 
 // RunFuncE delegates to stdinFunc if --stdin flag is passed otherwise delegates to promptFunc.
 func RunFuncE(stdinFunc, promptFunc CommandRunnerFunc) CommandRunnerFunc {
