+++ pkg/cmd/formula.go
@@ -31,7 +31,6 @@ import (
 	"github.com/ZupIT/ritchie-cli/pkg/api"
 	"github.com/ZupIT/ritchie-cli/pkg/formula"
 	"github.com/ZupIT/ritchie-cli/pkg/formula/input"
-	"github.com/ZupIT/ritchie-cli/pkg/prompt"
 	"github.com/ZupIT/ritchie-cli/pkg/slice/sliceutil"
 	"github.com/ZupIT/ritchie-cli/pkg/stream"
 )
@@ -74,16 +73,6 @@ var (
 		},
 	}
 )
-
-type flag struct {
-	name        string
-	shortName   string
-	kind        reflect.Kind
-	defValue    interface{}
-	description string
-}
-
-type flags []flag
 
 var ErrRunFormulaWithTwoFlag = errors.New("you cannot run formula with --docker and --local flags together")
 
@@ -218,22 +207,6 @@ func (f FormulaCommand) execFormulaFunc(repo, path string) func(cmd *cobra.Comma
 		}
 
 		return nil
-	}
-}
-
-func addReservedFlags(flags *pflag.FlagSet) {
-	for _, flag := range reservedFlags {
-		switch flag.kind { //nolint:exhaustive
-		case reflect.String:
-			flags.StringP(flag.name, flag.shortName, flag.defValue.(string), flag.description)
-		case reflect.Bool:
-			flags.BoolP(flag.name, flag.shortName, flag.defValue.(bool), flag.description)
-		case reflect.Int:
-			flags.IntP(flag.name, flag.shortName, flag.defValue.(int), flag.description)
-		default:
-			warning := fmt.Sprintf("The %q type is not supported for the %q flag", flag.kind.String(), flag.name)
-			prompt.Warning(warning)
-		}
 	}
 }
